deleted the multiple copies of the same file
