+++ bin/README.md
@@ -1,0 +1,3 @@
+# MeetingManger
+AC12001 - Introduction To Data Structures And Algorithms Final Project
+Java Assignment
